split tests for models and api into separate files
